requirements.txt
vectordb2==0.1.3
spacy==3.6.0
lightning==2.0.5

chromadb==0.3.25
langchain==0.0.226
openai==0.27.6
openllm==0.1.20
tiktoken==0.4.0



